Cast temporary values to make sure they fit in uint-types

+++ core/src/rom_factory.cpp
@@ -58,11 +58,13 @@ std::unique_ptr<IRom> RomFactory::from_bytes(std::istream &bytestream) {
         h.prg_ram_size = 1; // For compatibility reasons, 0 ram means 1 ram.
     }
 
-    uint8_t mapper = (h.flags_6 & 0xF0u) >> 4u;
+    uint8_t mapper = static_cast<uint8_t>(h.flags_6 & 0xF0u) >> 4u;
     mapper |= h.flags_7 & 0xF0u;
 
-    const auto prg_rom_byte_count = static_cast<uint32_t>(h.prg_rom_size * 16 * 1024);
-    const auto chr_rom_byte_count = static_cast<uint32_t>(h.chr_rom_size * 8 * 1024);
+    const auto prg_rom_byte_count =
+            static_cast<uint32_t>(h.prg_rom_size * 16 * 1024);
+    const auto chr_rom_byte_count =
+            static_cast<uint32_t>(h.chr_rom_size * 8 * 1024);
 
     const uint32_t expected_rom_size =
             sizeof(INesHeader) + prg_rom_byte_count + chr_rom_byte_count;
